Hiding default values of required parameters.
When setting up clusters in normal mode and thereby configuring blockchains with CommandAddConfiguration.kt, it was confusing with default values also on the required parameters. By instead initiating them to null (and thereby also modify their types to be nullable), they are no longer shown in the help text.

For parameter height, the default value is sane, thus it was kept, and therefore it was changed to no longer be required.